file rename and more fixes
